backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
ua-parser==0.18.0
user-agents==2.2.0
